Move AggregateCommitmentService into its own module
