+++ core/src/lib.rs
@@ -11,6 +11,7 @@ pub mod banking_stage;
 pub mod broadcast_stage;
 pub mod cluster_info_vote_listener;
 pub mod commitment;
+pub mod commitment_service;
 mod deprecated;
 pub mod shred_fetch_stage;
 #[macro_use]
