+++ core/src/replay_stage.rs
@@ -6,7 +6,8 @@ use crate::{
     cluster_info::ClusterInfo,
     cluster_info_vote_listener::VoteTracker,
     cluster_slots::ClusterSlots,
-    commitment::{AggregateCommitmentService, BlockCommitmentCache, CommitmentAggregationData},
+    commitment::BlockCommitmentCache,
+    commitment_service::{AggregateCommitmentService, CommitmentAggregationData},
     consensus::{ComputedBankState, StakeLockout, SwitchForkDecision, Tower},
     fork_choice::{ForkChoice, SelectVoteAndResetForkResult},
     heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
